Handle click rate response display.
