+++ client/AdminUI/src/app/components/user-admin/aggregate-stats/aggregate-stats.component.html
@@ -50,11 +50,11 @@
     <table class="table">
         <tr>
             <td>Average click rate across all customers</td>
-            <td class="text-right">{{detail?.click_rate_across_all_customers}}</td>
+            <td class="text-right">{{click_rate}}</td>
         </tr>
         <tr>
             <td>Average time to click across all customers</td>
-            <td class="text-right">{{detail?.average_time_to_click_all_customers}}</td>
+            <td class="text-right">{{avg_time_to_click}}</td>
         </tr>
     </table>
 </div>
